requirements.txt
openai
langchain
unstructured
tiktoken
python-dotenv
pre-commit
pypdf
faiss-cpu